requirements.txt
PyQt5==5.15.11
pystray==0.19.5
Pillow>=10.0.0
